refactor: move upgrade banner
